+++ common/src/main/java/net/nimajnebec/autoelytra/mixin/PauseScreenMixin.java
@@ -1,7 +1,7 @@
 package net.nimajnebec.autoelytra.mixin;
 
 import net.minecraft.client.gui.screens.PauseScreen;
-import net.nimajnebec.autoelytra.features.AutoEquipController;
+import net.nimajnebec.autoelytra.feature.AutoEquipController;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
